test: Verify API intergration
