+++ package.json
@@ -2,6 +2,7 @@
   "name": "unitalk",
   "version": "0.1.0",
   "private": true,
+  "main": "src/index.tsx",
   "dependencies": {
     "@date-io/date-fns": "^3.0.0",
     "@emailjs/browser": "^4.3.3",
@@ -28,17 +29,20 @@
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
     "react-redux": "^9.1.2",
-    "react-router-dom": "^6.24.0",
-    "react-scripts": "5.0.1",
+    "react-router-dom": "^6.24.1",
+    "react-scripts": "^3.0.1",
+    "recoil": "^0.7.7",
     "redux": "^5.0.1",
-    "web-vitals": "^2.1.4"
+    "tsconfig": "^7.0.0",
+    "web-vitals": "^2.1.4",
+    "webpack": "^4.29.6"
   },
   "scripts": {
     "start": "react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test",
     "eject": "react-scripts eject",
-    "stop": "'Stopping React development server'"
+    "stop": "echo 'Stopping React development server'"
   },
   "eslintConfig": {
     "extends": [
